ml_model/routes/auth.py
"""
Authentication Routes
Google OAuth and JWT token management
"""
from flask import Blueprint, request, jsonify
from datetime import datetime
import jwt
import os

from utils.logger import setup_logger
from utils.errors import AuthenticationException
from config.settings import active_config

logger = setup_logger(__name__)

bp = Blueprint('auth', __name__, url_prefix='/api/auth')

@bp.route('/google', methods=['POST'])
def google_oauth():
    """
    Handle Google OAuth token exchange
    
    TODO: Implement Google token verification
    For now, this is a placeholder that will:
    1. Accept Google ID token
    2. Verify token signature
    3. Create JWT for app use
    4. Return profile info
    
    Request Body:
        - token: str (Google ID token)
    
    Returns:
        - jwt_token: str  
        - user_info: dict (name, email, picture)
    """
    try:
        data = request.get_json()
        
        if not data or 'token' not in data:
            raise AuthenticationException("Missing 'token' field")
        
        google_token = data['token']
        
        # TODO: Verify Google token
        # For now, we'll return a placeholder response
        logger.info(f"Received Google OAuth token (length: {len(google_token)})")
        
        # In production, you would:
        # 1. Verify token with Google
        # 2. Extract user info
        # 3. Create JWT token
        
        # Placeholder response
        return jsonify({
            'status': 'success',
            'message': 'OAuth endpoint ready for integration',
            'note': 'Add Google API credentials to .env file',
            'required_fields': {
                'GOOGLE_CLIENT_ID': 'Your Google OAuth Client ID',
                'GOOGLE_CLIENT_SECRET': 'Your Google OAuth Client Secret'
            }
        }), 200
        
    except AuthenticationException as e:
        return jsonify({
            'status': 'error',
            'message': e.message
        }), 401
    except Exception as e:
        logger.error(f"OAuth error: {str(e)}")
        return jsonify({
            'status': 'error',
            'message': 'Authentication failed'
        }), 500

@bp.route('/verify-token', methods=['POST'])
def verify_token():
    """
    Verify JWT token validity
    
    Headers:
        - Authorization: Bearer <token>
    """
    try:
        auth_header = request.headers.get('Authorization')
        
        if not auth_header or not auth_header.startswith('Bearer '):
            raise AuthenticationException("Missing or invalid Authorization header")
        
        token = auth_header[7:]  # Remove 'Bearer '
        
        # Verify token
        payload = jwt.decode(
            token,
            active_config.JWT_SECRET,
            algorithms=[active_config.JWT_ALGORITHM]
        )
        
        return jsonify({
            'status': 'success',
            'valid': True,
            'user_id': payload.get('user_id'),
            'expires_at': payload.get('exp')
        }), 200
        
    except jwt.ExpiredSignatureError:
        return jsonify({
            'status': 'error',
            'message': 'Token expired'
        }), 401
    except jwt.InvalidTokenError:
        return jsonify({
            'status': 'error',
            'message': 'Invalid token'
        }), 401
    except Exception as e:
        return jsonify({
            'status': 'error',
            'message': str(e)
        }), 500

@bp.route('/refresh-token', methods=['POST'])
def refresh_token():
    """
    Refresh JWT token
    
    Request Body:
        - refresh_token: str
    """
    return jsonify({
        'status': 'success',
        'message': 'Token refresh endpoint (implement as needed)'
    }), 200
