ml_model/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
requests>=2.31.0
python-dateutil>=2.8.2
pytest>=7.4.0
black>=23.9.0
flake8>=6.0.0

